pyproject.toml
[project]
name = "data-platform"
version = "0.1.0"
requires-python = ">3.7"
readme = "README.md"
authors = [{ name = "Data Team", email = "data@team.com" }]

dependencies = [
  # Formatting
  "black",
  # Type checking
  "mypy",
  # Linting
  "ruff",
  # phidata
  "phidata==2.0.0.dev29",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# Exclude secrets from the package
# https://setuptools.pypa.io/en/latest/userguide/datafiles.html#exclude-package-data
[tool.setuptools.exclude-package-data]
workspace = ["secrets"]

# Change this value to use a different directory for the phidata workspace.
# [tool.phidata]
# workspace = "workspace"

[tool.ruff]
line-length = 110
exclude = [
  "dpenv*",
  "workspace/dev/airflow/config/*",
  "workspace/prd/airflow/config/*",
  "workspace/dev/jupyter/config/*",
  "workspace/prd/jupyter/config/*",
  "workspace/prd/superset/config/*",
]

[tool.mypy]
pretty = true
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = [
  "dpenv*",
  "workspace/dev/airflow/config/*",
  "workspace/prd/airflow/config/*",
  "workspace/dev/jupyter/config/*",
  "workspace/prd/jupyter/config/*",
  "workspace/prd/superset/config/*",
]

[[tool.mypy.overrides]]
module = ["airflow.*", "phi.*", "pyarrow.*", "setuptools.*"]
ignore_missing_imports = true
